more code review changes
