+++ js/main.js
@@ -1,6 +1,6 @@
 import { buildPostsData } from './data.js';
-import { renderGallery } from './gallery/gallery.js';
-import { renderUpload } from './upload/upload.js';
+import { renderGallery } from './gallery/main.js';
+import { renderUpload } from './upload/main.js';
 
 renderGallery(buildPostsData(19));
 renderUpload();
