+++ js/upload/popup.js
@@ -1,10 +1,8 @@
 import {isEscapeKey} from '../utilities.js';
-import {validateForm, resetForm} from './validation.js';
 
-const uploadForm = document.querySelector('.img-upload__form');
-const inputPhoto = uploadForm.querySelector('.img-upload__input');
-const editForm = uploadForm.querySelector('.img-upload__overlay');
-const closeButton = uploadForm.querySelector('.img-upload__cancel');
+const form = document.querySelector('.img-upload__form');
+const popup = form.querySelector('.img-upload__overlay');
+const closeButton = form.querySelector('.img-upload__cancel');
 
 const onUploadButtonClick = (evt) => {
   if (isEscapeKey(evt)) {
@@ -14,32 +12,16 @@ const onUploadButtonClick = (evt) => {
 };
 
 const closePopup = () => {
-  editForm.classList.add('hidden');
+  popup.classList.add('hidden');
   document.body.classList.remove('modal-open');
   document.removeEventListener('keydown', onUploadButtonClick);
-  uploadForm.reset();
 };
 
 const openPopup = () => {
-  editForm.classList.remove('hidden');
+  popup.classList.remove('hidden');
   document.body.classList.add('modal-open');
   closeButton.addEventListener('click', closePopup);
   document.addEventListener('keydown', onUploadButtonClick);
 };
 
-const renderPopup = () => {
-  inputPhoto.addEventListener('change', () => {
-    openPopup();
-  });
-
-  uploadForm.addEventListener('submit', (event) => {
-    event.preventDefault();
-    validateForm();
-  });
-
-  uploadForm.addEventListener('reset', () => {
-    resetForm();
-  });
-};
-
-export {renderPopup};
+export {openPopup};
