--- js/upload/upload.js
@@ -1,7 +1,0 @@
-import {renderPopup} from './popup.js';
-
-const renderUpload = () => {
-  renderPopup();
-};
-
-export {renderUpload};
